PROX-367: bump midgard-proto

+++ src/main/java/com/rbkmoney/midgard/utils/MappingUtils.java
@@ -77,8 +77,8 @@ public final class MappingUtils {
 
     private static TransactionCardInfo getTransactionCardInfo(ClearingTransaction clrTran) {
         BankCardExpDate expDate = new BankCardExpDate(
-                clrTran.getPayerBankCardExpiredDateMonth(),
-                clrTran.getPayerBankCardExpiredDateYear()
+                Byte.valueOf(clrTran.getPayerBankCardExpiredDateMonth()),
+                Short.valueOf(clrTran.getPayerBankCardExpiredDateYear())
         );
         return new TransactionCardInfo()
                 .setPayerBankCardToken(clrTran.getPayerBankCardToken())
